+++ css-inline/src/html/selectors/parser.rs
@@ -1,0 +1,46 @@
+use super::{pseudo_classes::PseudoClass, selector_impl::InlinerSelectors, ParseError};
+use cssparser::{CowRcStr, SourceLocation};
+use selectors::{parser::SelectorParseErrorKind, Parser};
+
+/// CSS selector parser.
+pub(crate) struct SelectorParser;
+
+impl<'i> Parser<'i> for SelectorParser {
+    type Impl = InlinerSelectors;
+    type Error = SelectorParseErrorKind<'i>;
+
+    fn parse_non_ts_pseudo_class(
+        &self,
+        location: SourceLocation,
+        name: CowRcStr<'i>,
+    ) -> Result<PseudoClass, ParseError<'i>> {
+        use self::PseudoClass::*;
+        if name.eq_ignore_ascii_case("any-link") {
+            Ok(AnyLink)
+        } else if name.eq_ignore_ascii_case("link") {
+            Ok(Link)
+        } else if name.eq_ignore_ascii_case("visited") {
+            Ok(Visited)
+        } else if name.eq_ignore_ascii_case("active") {
+            Ok(Active)
+        } else if name.eq_ignore_ascii_case("focus") {
+            Ok(Focus)
+        } else if name.eq_ignore_ascii_case("hover") {
+            Ok(Hover)
+        } else if name.eq_ignore_ascii_case("enabled") {
+            Ok(Enabled)
+        } else if name.eq_ignore_ascii_case("disabled") {
+            Ok(Disabled)
+        } else if name.eq_ignore_ascii_case("checked") {
+            Ok(Checked)
+        } else if name.eq_ignore_ascii_case("indeterminate") {
+            Ok(Indeterminate)
+        } else {
+            Err(
+                location.new_custom_error(SelectorParseErrorKind::UnsupportedPseudoClassOrElement(
+                    name,
+                )),
+            )
+        }
+    }
+}